Create README.md
This Wordbook offer interactive ways of finding the meanings of word in comparison to
printed dictionaries. In comparison to printed dictionaries, searching for word meanings
is only a simple click of the mouse on the online interface and scrolling up and down the
pages on the screen. Hence the meaning is found instantly. This is in contrast to the laborious efforts of flipping through the pages that hinder
learners to search for meaning. Its ease and fun to use feature is one of the main
motivators for students to use wordbook. The visual impacts foster interactive learning
and aid in enriching student’s vocabulary. Many users have found Wordbook useful in their daily life. The convenience of not
reading every single line is much preferred, and often times users don't have a good idea
of how to spell the word they are looking for so text-to-speech is much convenient for
them.